Changed the way timestamps are read from the database, so they are not converted from binary to string then back to binary again. Now we just convert binary to binary.
Updated internal timstamp storage in nodes, ways, relations to be quint64, instead of uint32... so we get ahead of that pesky 2038 problem.

Updated documentation for the map-diff command.

+++ hoot-core/src/main/cpp/hoot/core/elements/ElementData.h
@@ -57,7 +57,7 @@ public:
 
   static long CHANGESET_EMPTY;
   static long VERSION_EMPTY;
-  static unsigned int TIMESTAMP_EMPTY;
+  static quint64 TIMESTAMP_EMPTY;
   static QString USER_EMPTY;
   static long UID_EMPTY;
   static bool VISIBLE_EMPTY;
@@ -109,7 +109,7 @@ protected:
   ElementData(long id = LLONG_MIN, const Tags& tags = Tags(), Meters circularError = -1,
               long changeset = ElementData::CHANGESET_EMPTY,
               long version = ElementData::VERSION_EMPTY,
-              unsigned int timestamp = ElementData::TIMESTAMP_EMPTY,
+              quint64 timestamp = ElementData::TIMESTAMP_EMPTY,
               QString user = ElementData::USER_EMPTY, long uid = ElementData::UID_EMPTY,
               bool visible = ElementData::VISIBLE_EMPTY);
 
@@ -118,7 +118,7 @@ protected:
   Meters _circularError;
   long _changeset;
   long _version;
-  unsigned int _timestamp;
+  quint64 _timestamp;
   QString _user;
   long _uid;
   bool _visible;
@@ -126,7 +126,7 @@ protected:
 };
 
 inline ElementData::ElementData(long id, const Tags& tags, Meters circularError, long changeset,
-                         long version, unsigned int timestamp, QString user, long uid,
+                         long version, quint64 timestamp, QString user, long uid,
                          bool visible)
   : _id(id),
     _tags(tags),
